feat: extract column mapping to dedicated service
Extract column mapping logic (~110 lines) from frontend.rs into new
mapping/columns.rs module with ColumnMapper service. Frontend now uses
ColumnMapper::new(&encrypt) to handle projection, parameter, and literal
column mappings.

- Created mapping/columns.rs with ColumnMapper service
- Extracted get_projection_columns, get_param_columns, get_literal_columns, get_column methods
- Updated frontend.rs to use ColumnMapper service instead of direct methods
- Added proper imports for Column and EqlTerm types
- Cleaned up unused imports

Part of incremental frontend refactoring (Step 2/7).
