+++ packages/cipherstash-proxy/src/postgresql/frontend.rs
@@ -1,5 +1,6 @@
 use super::context::{Context, Statement};
 use super::error_handler::PostgreSqlErrorHandler;
+use super::mapping::ColumnMapper;
 use super::messages::bind::Bind;
 use super::messages::describe::Describe;
 use super::messages::execute::Execute;
@@ -10,11 +11,9 @@ use super::protocol::{self};
 use super::sql::SqlProcessor;
 use crate::connect::Sender;
 use crate::encrypt::Encrypt;
-use crate::eql::Identifier;
 use crate::error::{EncryptError, Error, MappingError};
 use crate::log::{CONTEXT, MAPPER, PROTOCOL};
-use crate::postgresql::context::column::Column;
-use crate::postgresql::context::Portal;
+use crate::postgresql::context::{column::Column, Portal};
 use crate::postgresql::data::literal_from_sql;
 use crate::postgresql::messages::close::Close;
 use crate::postgresql::messages::ready_for_query::ReadyForQuery;
@@ -29,10 +28,9 @@ use crate::prometheus::{
 use crate::EqlEncrypted;
 use bytes::BytesMut;
 use cipherstash_client::encryption::Plaintext;
-use eql_mapper::{self, EqlMapperError, EqlTerm, TableColumn, TypeCheckedStatement};
+use eql_mapper::{self, EqlMapperError, EqlTerm, TypeCheckedStatement};
 use metrics::{counter, histogram};
 use pg_escape::quote_literal;
-use postgres_types::Type;
 use serde::Serialize;
 use sqltk::parser::ast::{self, Value};
 use sqltk::NodeKey;
@@ -777,9 +775,10 @@ where
         typed_statement: &TypeCheckedStatement<'_>,
         param_types: Vec<i32>,
     ) -> Result<Option<Statement>, Error> {
-        let param_columns = self.get_param_columns(typed_statement)?;
-        let projection_columns = self.get_projection_columns(typed_statement)?;
-        let literal_columns = self.get_literal_columns(typed_statement)?;
+        let column_mapper = ColumnMapper::new(&self.encrypt);
+        let param_columns = column_mapper.get_param_columns(typed_statement)?;
+        let projection_columns = column_mapper.get_projection_columns(typed_statement)?;
+        let literal_columns = column_mapper.get_literal_columns(typed_statement)?;
 
         let no_encrypted_param_columns = param_columns.iter().all(|c| c.is_none());
         let no_encrypted_projection_columns = projection_columns.iter().all(|c| c.is_none());
@@ -981,157 +980,6 @@ where
         }
     }
 
-    ///
-    /// Maps typed statement projection columns to an Encrypt column configuration
-    ///
-    /// The returned `Vec` is of `Option<Column>` because the Projection columns are a mix of native and EQL types.
-    /// Only EQL colunms will have a configuration. Native types are always None.
-    ///
-    /// Preserves the ordering and semantics of the projection to reduce the complexity of positional encryption.
-    ///
-    fn get_projection_columns(
-        &self,
-        typed_statement: &eql_mapper::TypeCheckedStatement<'_>,
-    ) -> Result<Vec<Option<Column>>, Error> {
-        let mut projection_columns = vec![];
-
-        for col in typed_statement.projection.columns() {
-            let eql_mapper::ProjectionColumn { ty, .. } = col;
-            let configured_column = match &**ty {
-                eql_mapper::Type::Value(eql_mapper::Value::Eql(eql_term)) => {
-                    let TableColumn { table, column } = eql_term.table_column();
-                    let identifier: Identifier = Identifier::from((table, column));
-
-                    debug!(
-                        target: MAPPER,
-                        client_id = self.context.client_id,
-                        msg = "Configured column",
-                        column = ?identifier,
-                        ?eql_term,
-                    );
-                    self.get_column(identifier, eql_term)?
-                }
-                _ => None,
-            };
-            projection_columns.push(configured_column)
-        }
-
-        Ok(projection_columns)
-    }
-
-    ///
-    /// Maps typed statement param columns to an Encrypt column configuration
-    ///
-    /// The returned `Vec` is of `Option<Column>` because the Param columns are a mix of native and EQL types.
-    /// Only EQL colunms will have a configuration. Native types are always None.
-    ///
-    /// Preserves the ordering and semantics of the projection to reduce the complexity of positional encryption.
-    ///
-    ///
-    fn get_param_columns(
-        &self,
-        typed_statement: &eql_mapper::TypeCheckedStatement<'_>,
-    ) -> Result<Vec<Option<Column>>, Error> {
-        let mut param_columns = vec![];
-
-        for param in typed_statement.params.iter() {
-            let configured_column = match param {
-                (_, eql_mapper::Value::Eql(eql_term)) => {
-                    let TableColumn { table, column } = eql_term.table_column();
-                    let identifier = Identifier::from((table, column));
-
-                    debug!(
-                        target: MAPPER,
-                        client_id = self.context.client_id,
-                        msg = "Encrypted parameter",
-                        column = ?identifier,
-                        ?eql_term,
-                    );
-
-                    self.get_column(identifier, eql_term)?
-                }
-                _ => None,
-            };
-            param_columns.push(configured_column);
-        }
-
-        Ok(param_columns)
-    }
-
-    fn get_literal_columns(
-        &self,
-        typed_statement: &eql_mapper::TypeCheckedStatement<'_>,
-    ) -> Result<Vec<Option<Column>>, Error> {
-        let mut literal_columns = vec![];
-
-        for (eql_term, _) in typed_statement.literals.iter() {
-            let TableColumn { table, column } = eql_term.table_column();
-            let identifier = Identifier::from((table, column));
-
-            debug!(
-                target: MAPPER,
-                client_id = self.context.client_id,
-                msg = "Encrypted literal",
-                column = ?identifier,
-                ?eql_term,
-            );
-            let col = self.get_column(identifier, eql_term)?;
-            if col.is_some() {
-                literal_columns.push(col);
-            }
-        }
-
-        Ok(literal_columns)
-    }
-
-    ///
-    /// Get the column configuration for the Identifier
-    /// Returns `EncryptError::UnknownColumn` if configuration cannot be found for the Identified column
-    /// if mapping enabled, and None if mapping is disabled. It'll log a warning either way.
-    fn get_column(
-        &self,
-        identifier: Identifier,
-        eql_term: &EqlTerm,
-    ) -> Result<Option<Column>, Error> {
-        match self.encrypt.get_column_config(&identifier) {
-            Some(config) => {
-                debug!(
-                    target: MAPPER,
-                    client_id = self.context.client_id,
-                    msg = "Configured column",
-                    column = ?identifier
-                );
-
-                // IndexTerm::SteVecSelector
-                let postgres_type = if matches!(eql_term, EqlTerm::JsonPath(_)) {
-                    Some(Type::JSONPATH)
-                } else {
-                    None
-                };
-
-                let eql_term = eql_term.variant();
-                Ok(Some(Column::new(
-                    identifier,
-                    config,
-                    postgres_type,
-                    eql_term,
-                )))
-            }
-            None => {
-                warn!(
-                    target: MAPPER,
-                    client_id = self.context.client_id,
-                    msg = "Configured column not found. Encryption configuration may have been deleted.",
-                    ?identifier,
-                );
-                Err(EncryptError::UnknownColumn {
-                    table: identifier.table.to_owned(),
-                    column: identifier.column.to_owned(),
-                }
-                .into())
-            }
-        }
-    }
 
     ///
     /// Send an ReadyForQuery to the client and remove error state.
