+++ packages/cipherstash-proxy/src/postgresql/mapping/mod.rs
@@ -1,0 +1,3 @@
+pub mod columns;
+
+pub use columns::ColumnMapper;
